Move TimeComponent to di
