+++ component/__init__.py
@@ -1,0 +1,4 @@
+from typing import Tuple
+
+# x, y, w, h
+Box = Tuple[int, int, int, int]
