+++ di.py
@@ -2,6 +2,7 @@ from dependency_injector import containers, providers
 from data.purpleair import PurpleAirDataResolver
 from data.envcanada import EnvironmentCanadaDataResolver
 from data.calendar import CalendarDataResolver
+from component.time import TimeComponent
 from pixelperfectpi import Clock
 
 import pytz
@@ -24,9 +25,17 @@ class Container(containers.DeclarativeContainer):
         display_tz=display_tz,
     )
 
+    time_component = providers.Singleton(
+        TimeComponent,
+        box=(29, 0, 35, 13),
+        # , **addt_config)
+        font_path=config.font_path,
+    )
+
     clock = providers.Singleton(
         Clock,
         purpleair=purpleair,
         env_canada=env_canada,
         calendar=calendar,
+        time_component=time_component,
     )
