+++ draw/__init__.py
@@ -1,4 +1,5 @@
-from typing import Tuple
+from .drawpanel import DrawPanel
+from .multipanelpanel import MultiPanelPanel
+from .box import Box
 
-# x, y, w, h
-Box = Tuple[int, int, int, int]
+__all__ = [ 'DrawPanel', 'MultiPanelPanel', 'Box' ]
